fix(evpn): split netlink files
